refactor: jwt 폴더를 auth 하위 폴더로 설정
