+++ src/main/java/com/choogoomoneyna/choogoomoneyna_be/auth/jwt/JwtAuthenticationEntryPoint.java
@@ -1,4 +1,4 @@
-package com.choogoomoneyna.choogoomoneyna_be.jwt;
+package com.choogoomoneyna.choogoomoneyna_be.auth.jwt;
 
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
