+++ src/main/java/com/choogoomoneyna/choogoomoneyna_be/user/controller/UserController.java
@@ -1,13 +1,11 @@
 package com.choogoomoneyna.choogoomoneyna_be.user.controller;
 
-import com.choogoomoneyna.choogoomoneyna_be.jwt.JwtTokenProvider;
-import com.choogoomoneyna.choogoomoneyna_be.jwt.service.RefreshTokenService;
-import com.choogoomoneyna.choogoomoneyna_be.jwt.service.RefreshTokenServiceImpl;
+import com.choogoomoneyna.choogoomoneyna_be.auth.jwt.JwtTokenProvider;
+import com.choogoomoneyna.choogoomoneyna_be.auth.jwt.service.RefreshTokenService;
 import com.choogoomoneyna.choogoomoneyna_be.user.dto.LoginType;
 import com.choogoomoneyna.choogoomoneyna_be.user.dto.request.JwtTokenResponseDTO;
 import com.choogoomoneyna.choogoomoneyna_be.user.dto.request.UserJoinRequestDTO;
 import com.choogoomoneyna.choogoomoneyna_be.user.dto.request.UserLoginRequestDTO;
-import com.choogoomoneyna.choogoomoneyna_be.user.mapper.UserMapper;
 import com.choogoomoneyna.choogoomoneyna_be.user.service.UserService;
 import com.choogoomoneyna.choogoomoneyna_be.user.vo.UserVO;
 import lombok.RequiredArgsConstructor;
@@ -16,9 +14,7 @@ import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
