+++ src/main/resources/com/choogoomoneyna/refreshToken/mapper/RefreshTokenMapper.xml
@@ -3,9 +3,9 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.choogoomoneyna.choogoomoneyna_be.jwt.mapper.RefreshTokenMapper">
+<mapper namespace="com.choogoomoneyna.choogoomoneyna_be.auth.jwt.mapper.RefreshTokenMapper">
 
-    <resultMap id="refreshTokenResultMap" type="com.choogoomoneyna.choogoomoneyna_be.jwt.vo.RefreshTokenVO">
+    <resultMap id="refreshTokenResultMap" type="com.choogoomoneyna.choogoomoneyna_be.auth.jwt.vo.RefreshTokenVO">
         <id property="id" column="token_id"/>
         <result property="userId" column="user_id"/>
         <result property="refreshToken" column="refresh_token"/>
@@ -13,7 +13,7 @@
         <result property="expiresAt" column="expires_at"/>
     </resultMap>
 
-    <insert id="insertRefreshToken" parameterType="com.choogoomoneyna.choogoomoneyna_be.jwt.vo.RefreshTokenVO">
+    <insert id="insertRefreshToken" parameterType="com.choogoomoneyna.choogoomoneyna_be.auth.jwt.vo.RefreshTokenVO">
         INSERT INTO refresh_tokens (user_id, refresh_token, issued_at, expires_at)
         VALUES (#{userId}, #{refreshToken}, #{issuedAt}, #{expiresAt})
     </insert>
